// frontend/src/app/api/model/weightHistoryEntry.ts
/**
 * Life Management 2.0 API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface WeightHistoryEntry { 
    id: string;
    recordedAt: string;
    weightKg: number;
    deleted: boolean;
    readonly deletedAt?: string | null;
    readonly createdAt?: string;
    readonly updatedAt?: string;
}